Restructure protobuf generation
