+++ benchmark/read/read_benchmark.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/m3db/m3coordinator/benchmark/common"
-	"github.com/m3db/m3coordinator/generated/proto/prometheus/prompb"
+	"github.com/m3db/m3coordinator/generated/proto/prompb"
 	"github.com/m3db/m3coordinator/services/m3coordinator/config"
 
 	"github.com/m3db/m3db/client"
