+++ services/m3coordinator/handler/read_test.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"github.com/m3db/m3coordinator/executor"
-	"github.com/m3db/m3coordinator/generated/proto/prometheus/prompb"
+	"github.com/m3db/m3coordinator/generated/proto/prompb"
 	"github.com/m3db/m3coordinator/mocks"
 	"github.com/m3db/m3coordinator/policy/resolver"
 	"github.com/m3db/m3coordinator/storage"
