+++ services/m3coordinator/handler/write.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/m3db/m3coordinator/generated/proto/prometheus/prompb"
+	"github.com/m3db/m3coordinator/generated/proto/prompb"
 	"github.com/m3db/m3coordinator/storage"
 	"github.com/m3db/m3coordinator/util/execution"
 	"github.com/m3db/m3coordinator/util/logging"
