+++ tsdb/remote/client.go
@@ -5,7 +5,7 @@ import (
 	"io"
 
 	"github.com/m3db/m3coordinator/errors"
-	"github.com/m3db/m3coordinator/generated/proto/m3coordinator"
+	"github.com/m3db/m3coordinator/generated/proto/rpc"
 	"github.com/m3db/m3coordinator/storage"
 	"github.com/m3db/m3coordinator/ts"
 	"github.com/m3db/m3coordinator/util/logging"
