+++ tsdb/remote/codecs.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/m3db/m3coordinator/generated/proto/m3coordinator"
+	"github.com/m3db/m3coordinator/generated/proto/rpc"
 	"github.com/m3db/m3coordinator/models"
 	"github.com/m3db/m3coordinator/storage"
 	"github.com/m3db/m3coordinator/ts"
